Minimal socket and content support.
This is a pretty large commit (mea culpa) - but at least there's a
reasonable rationalization motivating it.  At its groundwork, we have a
few intertwined concepts, which we're representing here, but they all
have a central common link:

> We want to see a minimal response when browsing a remote public directory.

It followed that we needed at least basic support for understanding
socket input and output, as well as the ability to read directory
contents, individual file contents, and at least an informative response
if we hit something which wasn't there.

There's a few places in this house which need sweeping, but we're
punting on for the time being in the interests of publishing some new
code to master.  Most notably, we need better test coverage (and
probably better structure overall) to our concrete Responders.