Expand usage section in readme
We want a little more correct information in our README.  Previously it
wasn't specifying the `-jar` flag for its command line use, and there
was no elaboration on the purpose or name of the `-p` or `-d` jar flags.
This commit updates the documentation to be more reliable for those bits
of information.