Permit subdirectory listing navigation.
This is a pretty big commit, all things considered.  Why cut such a
large code change for a small update?  Mainly, test coverage.  In this
commit, you will find:

* Support for subdirectory navigation in directory listing pages.
* Massively expanded test coverage.
* A decoupling of output writing from responders.

This is an interstitial commit, overall - the goal here is to answer a
lot of the feedback provided for the first release of Duckling, as well
as to prepare for the next release (which will support routing).

Closes #6