Introduce .travis.yml file.
This commit wires the repository up to a Travis CI setup.  Ideally there
will be some jar releases which come along with things, but as of the
time of this commit we can't be sure.  You see, we have to push
something upstream before we can evaluate any of it.