Add stage task for Heroku support.
Heroku is kind of a big unknown here.  I can't imagine that the
directory listing, etc., will work - but it will be interesting to see
what happens.

This commit introduces a `stage` task to gradle, which is one of the
prerequisites for a Heroku deployment.  This deployment is currently set
to trigger immediately after a successful CI build on master.  In the
future that may change.