Allow route configuration.
Duckling now supports routing, although it doesn't expose configuration
just yet.  Because there's no external exposure of this configuration,
Duckling starts up with a batch of default routes instead.

This commit had four major goals.

1 - Decouple a lot of existing objects, in order to allow for better
test coverage.
2 - Add that better test coverage.
3 - Create a route definition system.
4 - Fulfill two more assertions in our Fitnesse acceptance test suite.

Because of those goals, there are large changes to the overall structure
of the server.  Most of these changes have cut down some on the
responsibility of a few objects (notably, Server) in exchange for a few
nested code packages.

One personal reservation I have with this update is that we've got a
good amount of similarly named packages, now.  I'd like to change that
as I get a feel for better names of things, but this commit represents a
worthwhile release candidate for now.