Omit bodies of HEAD requests.
In the case that we get a HEAD request for any route, we don't want to
publish a body with our response.

Under the hood this is instrumented as part of a larger refactor, where
given responders will know what sort of requests they allow.  This paves
the way for request handling to get a little more in depth as time goes
on.