Don't rely on overloaded equivalence checks.
Previously I was overloading the `#equals` method on a number of
objects, but if they'd ever been given an object which couldn't have
been coerced in the direction I assumed, that method would have resulted
in an exception and crash.

This commit updates those equivalence checks to assume false, except in
the event that the given object is actually a comparable object (in
other words, it's an instance of the same object).  In the case that the
offered object is comparable, we use the same comparison heuristics.