Add RouteContents to expand route abilities.
Routes previously only accepted strings, which they would use as bodies.
This commit prepares to overhaul that a bit, by adding an object which
can take a string or a (Request) -> String lambda.  In this way we hope
to empower the creation of more dynamic routes, instead of needing to
add a contents object for every unique route.