Support redirects on the route level.
This commit introduces redirect knowledge to the DefinedContents
object, all Response objects, and the Route object.  Basically, what
happens is that we treat a redirect as a normal content page, except we
add the content of the route to the headers list as a Location field.
This means that while we are rendering a string, most if not all clients
will point themselves towards the content as a destination URL.