Add some simple routes to the manifest.
Although things are starting to feel a little hairy as far as my object
graph is going, some things have gone very well.  In particular the
ability to write new simple routes is pretty easy.  This commit
introduces one new route that features two new methods, and updates the
README a little to talk about it.