Replace complex response lambda with Converter.
There was a Route definition which had a pretty complex internal logic
set.  This commit moves that into some intermediate objects which
conform to the same basic interface (`Function<Request, String>`).
These are intermediary, also in the sense that they represent a dabbling
into a potential replacement for the Responders pattern.