Expose MemoryCache functionality.
This commit introduces a new feature, MemoryCache access, and makes a
solid forward stride towards employing the new Routing approach with a
newly named pattern, Behaviors.

Behaviors are functional interfaces which can be used to compose
responses out of requests.  The truth is that they look a bit more like
middleware than a body composer, but they're a useful intermediary
anyway.