Introduce Response objects.
In order to prepare for a coming feature implementation, this commit
introduces Response objects and adopts them in each of our responders.
Where possible, we're going to steer route behavior, and default
behavior, towards using the semi-functional approach Responses can
offer.

To put it another way, we'd been creating some really complex objects
and this approach lets us instead model complicated things with a list
of functions that take a request and return a response.  Then, we merge
those responses.

Right now we've still got a bit of clean-up to accomplish, but this
release is the minimal functioning implementation.

Things left:

* Page function objects should be changed to Behavior objects.
* Routes should be changed to take a list of Behaviors.
* Responses should expose control of InputStreams to Behavior functions.

Some unanswered questions:

* Do we need responders at all anymore?  Can they go away?
* Can the Header / Header Pair structure go away?
* Is it possible to remove file knowledge from Requests?
* Can writers be simplified or removed?