Clean up pass, use behaviors for redirects.
Although it's kind of a big commit, the driving point behind this commit
was cleaning things up a bit.  We're going to rely on Route objects for
a whole lot less, for example - they no longer generate content or any
of that.  Really, they just map a list of behaviors to a method / route
combo now.

There was a bit of extra logic involved in merging responses, however,
so there is a new pair of objects in charge of that. MergeResponses
handles the merge itself, and ResponseBody wraps the string merge / body
wipe logic.  Eventually ResponseBody ought to help with ranges, as well.

Finally, it turned out that (although "clever"), the approach I was using to
compose responses was pretty slow.  Basically, I was composing responses
and rolling them up level by level.  A merge candidate would be composed
entirely before being merged into a target Response.

It was leading to tons of looping, but also it was creating a good
amount of strange bugs, and opening the application up to the
possibility of infinite loops if there were any self-referential
Behavior objects in a composition web.

The new approach to merging simply concatenates behavior lists, which is
faster, less complicated and coincidentally still does the trick
perfectly.