Allow partial content requests for files.
This commit introduces a new concept, which is the idea that we can
filter a response body before outputting it.  This concept works a
little bit like a behavior, but right now we're gatekeeping it behind
behaviors for the sake of consistency.

An overview of the sibling concepts which made it into this update:

* Responses and ResponseBodies now support streams.
* Responses can accept and track response body filters.
* Merging responses combines response body filters.
* FileContents responder now fully relies on behaviors.

From a bird's eye view, this and the coming options request change means
that we might be able to deprecate or meaningfully change responders.