Remove Logger.
Logging output is definitely an overall nice thing to have, but the
truth is that there weren't any stories which relied on it.  Secondly,
it turns out that the logging IO, which was coupled with Duckling's
socket loop, was having an unpredictable (often negative) impact on
overall response times.

This isn't the only optimization we'll need to make in order to optimize
duckling, but it's definitely the simplest.