Add Makefile and expanded build instructions.
In order to improve the quality of life of anyone who might take a
gander at duckling, we're going to add a Makefile with the most common
basic operations you might want to use gradle for.  The Makefile will
gracefully downgrade to ./gradlew if there isn't a gradle available on
the host machine.